Add underscore in front of internal method
Since all models inherit from ActiveRecord::Base, any private
method added by the framework can clash with existing methods
in the models defined in the app. Since this is an internal method
that was just introduced and has a change of clashing with methods
in app (I just found one), better to define using the `_`.