Resets the destroyed flag when reloading an object
There are cases where we have to rollback a deletion of a record
without being able to rely on a classic transaction block.

Calling reload on the destroyed (and then inserted) instance would
still mark the instance as being destroyed anymore because the
`@destroyed` flag is not getting resetted within the `reload` call.

closes #40820