Remove update_table_definition entry from migration log
6.0 `add_reference` compatibility migration method triggered not wanted
long message printed to stdout for `update_table_definition` invocation.
The reason for this was that `method_missing` method was invoked which
is responsible, among others, for printing migration logs and delegating
the invocation into `connector`. To remove unwanted log entry
`update_table_definition` is invoked directly on the `connector`
instance.