Fix failing tests introduced by #41030
In #41030 the tests didn't show as failing because they only fail when
run in isolation which happens on the merge to master.

In this test we need to re-establish the connection to
`ActiveRecord::Base` or else the tests get the connections confused and
for some reason some of the migration tests were looking for the wrong
connection specification name.