Fix CurrentAttributes.set when attribute readers are made private
As the documentation for CurrentAttributes notes, one must be careful
to limit the interface of a global singleton to avoid tangling up one’s
model. One strategy is to make a base attribute reader private and
delegate only the methods that should be exposed. For example,
CurrentRequest:

```ruby
class CurrentRequest < ActiveSupport::CurrentAttributes
  attribute :request
  private :request

  delegate :user_agent, :host, to: :request, allow_nil: true
end
```

Before this change, `CurrentRequest.set(request: request)` would fail
with ``NoMethodError: private method `request' called
for #<CurrentRequest:0x00007f8c96d85e80 @attributes={}>``. This is
because `compute_attributes` was using `public_send` to fetch the
current value for the attribute.

This change swaps `public_send` for `send`. Since `compute_attributes`
is a private method, it stands to reason that it should be able to call
other private methods.