AR connection configuration accept `url` as a key if `adapter` is set
Some ActiveRecord Adapters accept `url` as a configuration key.

This change passes `url` to the adapter if the adapter is also specified in the
configuration.

For example:

```yaml
production:
  adapter: sunstone
  url: 'https://example.com/'
````

Currently uses `url` to overwrite the adapter as "http" and host as "example.com".
This change would keep url and not overwrite the adapter.