Only override Github git_source for legacy Bundler
Bundler 1.x used an insecure `git://` URL as the default for gems
specified with the `:github` source in Gemfiles, and this common
workaround was added to the Gemfile template for newly generated apps to
override the then Bundler default.

As of Bundler 2.0.0, the `:github` source in Gemfiles defaults to HTTPS
(c.f. https://github.com/rubygems/bundler/releases/tag/v2.0.0) – so we
no longer need to include this if we generate a Gemfile in an
environment running a recent version of Bundler.

When Rails moves to require Bundler >= 2.0, this can be removed
entirely.