Fix `scope_for_create` to take only equality nodes
Related #39683.

If a scope has IN cluase, `scope_for_create` which is passed to
`assign_attributes` will include array values, and it will cause weird
behaviors.

That behavior was due to the `In` node was a subclass of the `Equality`
node (now the checking has migrated to `equality?` method).

i.e. It works unless IN queries, for example (BETWEEN):

```ruby
user = User.where(role: 0..1).find_or_initialize_by(name: "name")
```

I think that the `In` node was a subclass of the `Equality` node is
accidental, so the behavior is almost a bug to me.

This makes `scope_for_create` takes only equality nodes to fix the
issue.

Fixes #41295.