PostgreSQL: handle `timestamp with time zone` columns correctly in `schema.rb`
Found this issue while working on https://github.com/rails/rails/pull/41084

Currently if you have a `timestamp with time zone` column, and you run `rake db:schema:dump`, it will be output as a `t.datetime`. This is wrong because when you run `rake db:schema:load`, the column will be recreated as a `timestamp without time zone`.

The fix is to add a new column type, `t.timestamptz`. This behaves exactly the same as before, the only change is what native type it is converted to during schema loads.