Support phone_to options when name not specified
When `name` is not specified, `phone_to` uses `phone_number` as the link
text.  This commit allows options to be specified in such a case.  This
enables using the `country_code` option to include the country code in
the link `href`, while a country-code-less `phone_number` is used as the
link text.