Add plugin to common command USAGE
Because the PluginCommand is defined with `hide_command!`, running
`bin/rails -h` hides the plugin command.

As an alternative to removing the `hide_command!` this adds the plugin
command to the USAGE. Unless it's an engine, because adding an engine to
an engine doesn't make sense.

Both the plugin and engine terms are used so a user can search for both.