Fix schema dumper for infinite dates in PostgreSQL
A default value of Float::INFINITY for a date does not round-trip
correctly. It will be correctly loaded as 'infinity'::date, but
when the schema is dumped, it will appear in the schema file as
"Infinity," which then will not load correctly.

This change is based on the fixes for #22396 and #40751, which
address the exact same issue for floats and date-times, respectively.