Should not change serializable value for the enum attribute
`EnumType#serializable?` which allows values only in the enum mapping
was added in 72fd0ba, and it is called only in `HomogeneousIn`, it
caused the serialization inconsistency whether the values are in `IN`
clause or not.

```ruby
# SELECT "users".* FROM "users" WHERE "users"."role" = '0'
User.where(role: "0")

# SELECT "users".* FROM "users" WHERE "users"."role" = '1'
User.where(role: "1")

# SELECT "users".* FROM "users" WHERE "users"."role" IN (NULL)
User.where(role: ["0", "1"])
```

Even if we introduce validation for the enum serialization in the future,
we should at least not cause the inconsistency for now.

Fixes #41474.