Remove no-op `PredicateBuilder::BaseHandler` constant
That is to prevent the "uninitialized constant" error when loading the
data marshalized by the previous version of Rails (Rails 6.0), it is not
needed for the future version of Rails.

See #39611.

I've added the no-op constant in that time, but we usually don't
guarantee the marshalized object compatibility.

This also avoids the "instance variable @future_result not initialized"
warning in the `test_marshal_load_legacy_relation`.

https://buildkite.com/rails/rails/builds/75166#3d7e57df-3679-4fa8-9f17-e1b0533c3ccf/1074-1081