Revert "Merge pull request #39759 from kamipo/marshal_load_legacy_ar_object"
This reverts commit 4fe1452534540851461e99b3d0fecaf27931ab93, reversing
changes made to 0f4258f646a76869157a2f34120091d36f24f2b0.

That is to prevent the "uninitialized constant" error when loading the
data marshalized by the previous version of Rails (Rails 6.0), it is not
needed for the future version of Rails.

See #39759.

I've added the constant alias in that time, but we usually don't
guarantee the marshalized object compatibility.