Remove `finalize` from NullSession
`finalize_session` should not be able to be called if a session has not
been started, we should never have a `NullSession` when we get there. So
we can remove the guard altogether to help make it obvious if this sort
of thing does happen when it's not supposed to.