Allow preloader to batch at deeper levels
Previously we taught the preloader to batch similar queries together
when possible, however (to reduce scope) we only did this for preloads
at the top level of the provided Hash.

This commit adds a new private Associations::Preloader::Batch class
which now handles the batching operation and can do so for loaders from
different preloaders. This also modifies child_preloaders to return an
un-executed preloader instead of running immediately, so that we can
perform batching as we recurse each level.

One intentional limitation of this is that we're only attempting to
batch together associations which are at the same depth in the provided
Hash. This greatly simplifies the implementation and ensures a
consistent grouping happens.

Another limitation (which I hope to address later) is that this still
does not attempt to perform any batching of :through associations
(however children of through associations may still be batched).