Preserve index sort order when altering sqlite3 table
Any method in the SQLite3Adapter that calls #alter_table results in
the table losing sort order on its indexes. This is because the index.orders
option is not included in the options kwargs when the index(es) are added to the
copied table. We need to preserve the order option in #copy_table_indexes.

Co-authored-by: Étienne Barrié <etienne.barrie@gmail.com>