Fix preloading through_relation multiple times with different scopes
Previously when the same `through_relation` was preloaded multiple times but using
a different scope the relation was incorrectly cached; this led
into returning the wrong results.

The issue is fixed by checking if the relation to preload has a `preload_scope`
and is preloaded from a `through_relation`.

Fixes #41378