s/href/title/ to avoid URL _and_ HTML escaping
Since Nokogiri is responsible for transforming nodes into HTML, it's
strict handling of the `[href]` attribute is _URL_ escaping values,
which was not occurring in the prior tests.

In an effort to exercise the code in the same spirit as the original
coverage, replace the `href:` assignments with `title:` assignments,
since they're both attributes but a `title` has no restrictions on
format or validity.