Fix numericality validator on decimal attribute with implicit scale
#38212 added scale support to numericality validator on decimal
attribute.

That works as expected if a decimal attribute has explicit scale.

On the database-backed decimal type, implicit scale means that has no
scale. But the implementation of #38212 does no truncation in that case
(this behavior is correct for non-database-backed decimal type though).

To fix the issue on the database-backed decimal type, use the attribute
type object rather than extracting precision and scale and parsing
values on its own.