Use existing support for previous types
This uses the changes from https://github.com/rails/rails/pull/39929. Instead of doing a manual casting of
the previous value as part of the type, it leverages the existing support to do this now.

It also fixes a problem when loading a class with an encrypt declaration that wasn't backed
by a table was failing.

CC @kaspth