Introduces Action Text for API only application
Problem statement: Action Text cannot be used with the API-only
application as the action installer warns the missing webpacker
configurations and tries to install Action Text packages which
are not required in API-only app. Secondly, Action Text is best
solution for rich text content editing with support of embeds
and works well with Trix editor but has some missing documentation
which can be confusing at times.

This PR attempts to fix the Action Text installer and document
the usage. Action Text would still fail `post.content.body.to_html`
as it would return an empty ActionText::Content.
This is because `to_html` depends on `ApplicationController.renderer`
which is absent in API only application.

Related to #39266

Please feel free to close the PR if this change is not required. :)