Add Preloader::Association::LoaderQuery
This class represents the query a Preloader::Association uses to load
its records. This replaces the previous concept of a grouping_key and
serves both to group Preloader::Associations by their query and to run
the query to load the records (which nicely avoids using .first).

This should not change any functionality but should be slightly faster
due to being able to avoid calling `scope.to_sql` unless necessary.